pyproject.toml
[tool.poetry]
name = "useful"
version = "0.1.0"
description = ""
authors = ["Mikhail <mixail.critsyn@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.75.0"}
project_name = "^0.2.175"
emails = "^0.6"
PyJWT = "^2.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
psycopg2-binary = "^2.9.3"
Authlib = "^1.0.0"
httpx = "^0.22.0"
oauth = "^1.0.1"
base = "^1.0.4"
username = "^2.1.0"
typer = "^0.4.0"
social = "^0.1.dev3"
tortoise-orm = "^0.18.1"
aerich = "^0.6.2"
asyncpg = "^0.25.0"
password = "^0.2"
Avatar = "^0.1"
SQLAlchemy = "^1.4.32"
kwargs = "^1.0.1"

[tool.poetry.dev-dependencies]


[tool.aerich]
tortoise_orm = "src.config.tortoise_conf.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
